Fix(app): Resolve stuck loading screen and improve theme consistency
This commit addresses an issue where the application would get stuck on the entry screen. It also redesigns the loading screen to be more thematically consistent with the application's neon/cyberpunk aesthetic.

Key changes:
- **Bug Fix:** Refactored the main session board into a server component (`SessionBoard`) that fetches initial data. This correctly triggers the Suspense boundary in `app/page.tsx`, resolving the "stuck" screen issue.
- **Client Component Extraction:** Moved all client-side logic and state management from the old session board into a new `SessionClient` component.
- **Redesign:** Updated the `app/loading.tsx` component to use a "neon" themed spinner instead of the "matrix" variant, aligning it with the overall application theme.
- **Refactoring:** Created a shared mock Supabase client to facilitate testing and development without requiring real environment variables, which was necessary to get the frontend verification to pass.
