+++ app/loading.tsx
@@ -1,5 +1,5 @@
 import LoadingSpinner from "@/components/common/loading-spinner";
 
 export default function Loading() {
-  return <LoadingSpinner text="Entering the matrix..." variant="matrix" />;
+  return <LoadingSpinner text="Assembling the chaos..." variant="neon" />;
 }
